% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicSimulator.r
\name{makeTable}
\alias{makeTable}
\title{Detailed Report}
\usage{
makeTable(iScenario, res)
}
\arguments{
\item{iScenario}{number of scenario to be displayed}

\item{res}{result from a dynamic hedging simulation function, such as \code{\link{deltaHedge}}}
}
\description{
Builds a detailed report by time step as an xtable, with 6 columns:
\enumerate{
\item time step
\item stock price
\item delta
\item option value
\item bond position
\item portfolio value
}
}
