% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathSimulator.r
\name{rnormInnovations}
\alias{rnormInnovations}
\title{Generator of Random Matrices (rnorm)}
\usage{
rnormInnovations(nbPaths, pathLength)
}
\arguments{
\item{nbPaths}{[numeric] number of paths}

\item{pathLength}{[numeric] number of steps}
}
\value{
a matrix of uniform (0,1) random numbers
}
\description{
Random matrix generator
}
