\name{sobolInnovations}
\alias{sobolInnovations}
\title{Generator of Random Matrices (Sobol)}
\usage{
  sobolInnovations(nbPaths, pathLength)
}
\arguments{
  \item{nbPaths}{[numeric] number of paths}

  \item{pathLength}{[numeric] number of steps}
}
\value{
  a matrix of normal(0,1) random numbers
}
\description{
  Random matrix generator
}

