\name{sobolSeedInnovations}
\alias{sobolSeedInnovations}
\title{Generator of Random Matrices (Sobol, fixed seed)}
\usage{
  sobolSeedInnovations(nbPaths, pathLength)
}
\arguments{
  \item{nbPaths}{[numeric] number of paths}

  \item{pathLength}{[numeric] number of steps}
}
\value{
  a matrix of normal(0,1) random numbers
}
\description{
  Random matrix generator
}

