\name{rnormInnovations}
\alias{rnormInnovations}
\title{Generator of Random Matrices (rnorm)}
\usage{
  rnormInnovations(nbPaths, pathLength)
}
\arguments{
  \item{nbPaths}{[numeric] number of paths}

  \item{pathLength}{[numeric] number of steps}
}
\value{
  a matrix of normal(0,1) random numbers
}
\description{
  Random matrix generator
}

